libknot: change knot_edns_client_subnet_parse option length from size_t to uint16_t

+++ src/libknot/rrtype/opt.c
@@ -659,7 +659,7 @@ int knot_edns_client_subnet_write(uint8_t *option, size_t option_len,
 
 _public_
 int knot_edns_client_subnet_parse(knot_edns_client_subnet_t *ecs,
-                                  const uint8_t *option, size_t option_len)
+                                  const uint8_t *option, uint16_t option_len)
 {
 	if (ecs == NULL || option == NULL) {
 		return KNOT_EINVAL;
